Fix typo in comment (#4634)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

I am going to rewrite a comment that had typo

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

I am not going to test it at all 😝 
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Co-authored-by: Aleksandra Cynk <aleksandracynk@swmansion.com>