Bump react-native to 0.72.1 (#4662)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR bumps React Native in Example and FabricExample apps from
0.72.0-rc.6 to 0.72.1 as well as bumps react-native-gesture-handler to
2.12.0, react-native-svg to 13.9.0 and react-native-safe-area-context to
4.6.3.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->