Remove ErrorHandler (#4723)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently, `ErrorHandler` is referenced in 11 files and has 2 separate
implementations (one for Android and one for iOS) but it turns out that
since Reanimated v3 there's only 1 place where we use it:


https://github.com/software-mansion/react-native-reanimated/blob/ae2b9ded42ec03589e1fccd1ecc8ad3d967d67c2/android/src/main/cpp/NativeProxy.cpp#L532-L535

This PR removes `ErrorHandler` completely from our codebase in favor of
old simple `throw std::runtime_error`.

## Test plan

Check if CI is green.