Improve transform matrix animation (#4536)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
We used to animate transformation matrices the very same way we animate
every array - move each value from initial to final position. However
this may look weird in case of transformation matrices, since scale and
rotation may appear to change randomly - the solution is to decompose
actual transformation matrix into:

- translation
- skale
- rotation
  - rotation along X axis 
  - rotation along Y axis 
  - rotation along Z axis 
- skew (shear)

and then we animate translation, skale and skew the way we used to, but
for rotation we calculate all the three angles and get matrix for
particular angle. Finally we multiply all the matrices in correct order.

See the recording for comparison. The $\text{\large
\color{blue}{\color{lime}u}pper}$ square is animated with improved
transform matrix interpolation, the $\text{\large
\color{darkorange}{\color{red}l}ower}$ one uses the old technique.


https://github.com/software-mansion/react-native-reanimated/assets/56199675/974c983c-0948-4c75-a833-46b99e93e0d1



## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->
I've added a new test suite

![image](https://github.com/software-mansion/react-native-reanimated/assets/56199675/7e0aae7e-f008-411c-9c47-3d0b31b45f48)

---------

Co-authored-by: Aleksandra Cynk <aleksandracynk@swmansion.com>
Co-authored-by: Samuel Rodriguez-Lozano <44348413+samuel-rl@users.noreply.github.com>
Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>
Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>
Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>
Co-authored-by: Roland Yeghiazaryan <roland.yeghiazaryan@gmail.com>
Co-authored-by: mstach60161 <mstach60161@gmail.com>
Co-authored-by: Michał Mąka <62388446+michalmaka@users.noreply.github.com>
Co-authored-by: Michał Mąka <michalmaka@swmansion.com>