Rename `runtime_` to `rnRuntime_` in NativeProxy (#4765)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR renames ambiguous variable `runtime_` to `rnRuntime_` in
NativeProxy class on Android.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->