Move `JSRuntimeHelper` to a separate file (#4768)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves `JSRuntimeHelper` class (along with `CoreFunction` class)
to a separate file in order to decouple it from `Shareables.h`.
Eventually, we will unify `JSRuntimeHelper` and `RuntimeManager` into a
single class so this file will be deleted in the future.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->