Unify decorating RN runtime (#4773)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves duplicated logic of decorating RN runtime from
`NativeProxy.cpp` and `REAModule.mm` into
`RuntimeDecorator::decorateRNRuntime`. I've also tried to move
`__reanimatedModuleProxy` (`NativeReanimatedModule`) there but there
were some problems with circular headers which I couldn't solve easily
so I'm leaving this for the next PR.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->