Properly type `__reanimatedModuleProxy` (and other stuff) (#4780)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

The motivation behind this PR is that `__reanimatedModuleProxy` (which
is injected via JSI) has the same type as `NativeReanimated` class which
is a JS-only adapter that wraps JSI binding. While this allows us to
save a few lines of code, it breaks the mental model of adapters having
a potentially different API than wrapped instances.

This PR adds an explicit type `NativeReanimatedModule` type which is
used for `__reanimatedModuleProxy`. Apart from this, I've decoupled
`JSReanimated` from `NativeReanimated` by eliminating inheritance as
well as refactored the constructor of `NativeReanimated` to simplify the
logic behind it.

Finally, I've removed `updateDataSynchronously` method which not only
was unused but also didn't work at all (it's used only from the UI
runtime as `_updateDataSynchronously`).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->