DRY up build.gradle variables (#4792)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
not a big PR but noticing that some of the string literals are just the
variable on top so i'm DRYing up the string literals.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
build.gradle should still run correctly

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->