Make babel version detection more verbose (#4791)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently the error tells us that something is wrong but we don't have a
clue what could be causing it. This PR makes error more verbose,
displaying code of the offending worklet. I also added some cheeky
Example for that purpose.

## Test plan

Run new `BabelVersioningExample`

🚀

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>