Split `Scheduler` into `JSScheduler` and `UIScheduler` (#4800)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

💡 I recommend reviewing this PR commit-by-commit.

Currently, `Scheduler` class has two responsibilities &ndash; scheduling
stuff on the UI thread (platform-specific, written mostly in
Java/Obj-C++) and scheduling stuff on the JS thread (platform-agnostic,
written in C++). Additionally, importing `Scheduler.h` doesn't work on
iOS because the one from React Native takes precedence over ours. I
thought it's a great moment to finally split them up.

This PR extracts JS scheduler from `Scheduler` class and sprinkles the
code with some const-refs. Also, one might say `JSScheduler` doesn't
make sense since we already have `CallInvoker` but this layer of
abstraction comes in handy since it allows us to extend its
possibilities in the future (e.g. obtain the exclusive access to the
main RN runtime on the UI thread and execute the callback
synchronously).

In the next episode, we'll take a look on how to improve the Android
implementation of Scheduler which is barely understandable in its
current shape.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->