Restore `isConfigured` function (#4832)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

`isConfigured` was an internal check that due it's directory's
`index.ts` having `export * from ./core` was available as a direct
import from `react-native-reanimated`. Therefore we cannot simply remove
it in a minor version - deleting it raised
https://github.com/react-navigation/react-navigation/issues/11498,
`react-navigation` uses this function
[here](https://github.com/react-navigation/react-navigation/blob/24c03924397a6e59aba9f6b74a9c5cb4b939d9e1/packages/react-native-drawer-layout/src/views/Drawer.tsx#L56)
to detect Reanimated 3 since Reanimated 2 didn't have this function.

This PR restores this function, marks it as deprecated and adds a
separate `public` check to detect Reanimated 3.

## Test plan

🚀

## Note

We should probably locate more functions that are exported as a part of
public API but they shouldn't be.