Assert shareable type (#4869)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR replaces a few occurrences of `extractShareableOrThrow` with
`extractShareableOrThrow<T>` which asserts the type of the shareable
along with a proper error message.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->